+++ .github/workflows/ci_gradle.yml
@@ -69,7 +69,14 @@ jobs:
           spring.datasource.username: ${{ secrets.DB_USERNAME }}
           spring.datasource.password: ${{ secrets.DB_PW }}
 
-
+      - name: 5) Set application-s3.yml
+        uses: microsoft/variable-substitution@v1
+        with:
+          files: ./core/core-infra-s3/src/main/resources/application-db.yml
+        env:
+          spring.datasource.url: ${{ secrets.DB_URL }}
+          spring.datasource.username: ${{ secrets.DB_USERNAME }}
+          spring.datasource.password: ${{ secrets.DB_PW }}
 
       - name: Grant execute permission for gradlew and build
         run: |
