+++ api/build.gradle
@@ -2,6 +2,7 @@ dependencies {
 
     implementation project(':core:core-domain');
     implementation project(':core:core-infra-db');
+    implementation project(':core:core-infra-s3');
 
     // Core
     implementation 'org.springframework.boot:spring-boot-starter-web'
@@ -35,8 +36,8 @@ dependencies {
     // Validation
     implementation 'org.springframework.boot:spring-boot-starter-validation'
 
-    // S3
-    implementation 'org.springframework.cloud:spring-cloud-starter-aws:2.2.6.RELEASE'
+//    // S3
+//    implementation 'org.springframework.cloud:spring-cloud-starter-aws:2.2.6.RELEASE'
 
     // email
     implementation 'com.sun.mail:jakarta.mail:2.0.1'
