+++ api/src/main/java/com/sponus/sponusbe/domain/s3/S3TestController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.sponus.coredomain.domain.common.ApiResponse;
+import com.sponus.coreinfras3.S3Util;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,25 +20,25 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/api/v1/s3")
-public class S3Controller {
+public class S3TestController {
 
-	private final S3Service s3Service;
+	private final S3Util s3Util;
 
 	@PostMapping(value = "/uploadImage", consumes = "multipart/form-data")
 	public ApiResponse<String> uploadImage(@RequestPart(value = "file", required = false) MultipartFile file) {
-		return ApiResponse.onSuccess(s3Service.uploadFile(file));
+		return ApiResponse.onSuccess(s3Util.uploadFile(file));
 	}
 
 	@DeleteMapping(value = "/deleteImage")
 	public ApiResponse<String> deleteImage(@RequestBody String path) {
 		String image = path.substring(path.lastIndexOf('/') + 1);
-		return ApiResponse.onSuccess(s3Service.deleteImage(image));
+		return ApiResponse.onSuccess(s3Util.deleteFile(image));
 	}
 
 	@PostMapping(value = "/uploadImages", consumes = "multipart/form-data")
 	public ApiResponse<List<String>> uploadImages(
 		@RequestPart(value = "files", required = false) List<MultipartFile> files) {
-		return ApiResponse.onSuccess(s3Service.uploadFiles(files));
+		return ApiResponse.onSuccess(s3Util.uploadFiles(files));
 	}
 
 }
