+++ api/src/main/resources/application-prod.yml
@@ -27,18 +27,18 @@ spring:
       host: ${REDIS_HOST}
       port: ${REDIS_PORT}
 
-cloud:
-  aws:
-    s3:
-      bucket: ${BUCKET}
-      folder: images/
-    stack:
-      auto: false
-    region:
-      static: ${REGION}
-    credentials:
-      accessKey: ${S3_ACCESS_KEY}
-      secretKey: ${S3_SECRET_KEY}
+#cloud:
+#  aws:
+#    s3:
+#      bucket: ${BUCKET}
+#      folder: images/
+#    stack:
+#      auto: false
+#    region:
+#      static: ${REGION}
+#    credentials:
+#      accessKey: ${S3_ACCESS_KEY}
+#      secretKey: ${S3_SECRET_KEY}
 
 mail:
   smtp:
