+++ core/core-infra-s3/build.gradle
@@ -1,0 +1,7 @@
+dependencies {
+    // S3
+    implementation 'org.springframework.cloud:spring-cloud-starter-aws:2.2.6.RELEASE'
+}
+
+bootJar { enabled = false }
+jar { enabled = true }
