+++ core/core-infra-s3/src/main/java/com/sponus/coreinfras3/S3Config.java
@@ -1,4 +1,4 @@
-package com.sponus.sponusbe.domain.s3;
+package com.sponus.coreinfras3;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
