+++ core/core-infra-s3/src/main/resources/application-s3.yml
@@ -1,0 +1,13 @@
+# S3
+cloud:
+  aws:
+    s3:
+      bucket: ${BUCKET}
+      folder: images/
+    stack:
+      auto: false
+    region:
+      static: ${REGION}
+    credentials:
+      accessKey: ${S3_ACCESS_KEY}
+      secretKey: ${S3_SECRET_KEY}
