+++ settings.gradle
@@ -2,3 +2,4 @@ rootProject.name = 'sponus-be'
 include 'api'
 include ':core:core-domain'
 include ':core:core-infra-db'
+include ':core:core-infra-s3'
